% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CircuitInformation}
\alias{CircuitInformation}
\title{Circuit Information}
\format{
A data.frame with 3 columns
\describe{
 \item{Circuit}
 \item{Locality}
 \item{Country}
 \item{Season}
}
}
\source{
\url{http://ergast.com/mrd/methods/circuits/}
}
\usage{
CircuitInformation
}
\description{
Circuit Information
}
\keyword{datasets}
