% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConstructorInformation}
\alias{ConstructorInformation}
\title{Constructor Information}
\format{
A data.frame with 3 columns
\describe{
 \item{ConstructorName}
 \item{Nationality}
 \item{Season}
}
}
\source{
\url{http://ergast.com/mrd/methods/constructors/}
}
\usage{
ConstructorInformation
}
\description{
Constructor Information
}
\keyword{datasets}
