% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConstructorStandings}
\alias{ConstructorStandings}
\title{Constructor Standings}
\format{
A data.frame with 7 columns
\describe{
 \item{Pos}
 \item{Constructor}
 \item{Nationality}
 \item{Points}
 \item{Wins}
 \item{Season}
 \item{Round}
}
}
\source{
\url{http://ergast.com/mrd/methods/standings/}
}
\usage{
ConstructorStandings
}
\description{
Constructor Standings
}
\keyword{datasets}
