% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DriverInformation}
\alias{DriverInformation}
\title{Driver Information}
\format{
A data.frame with 5 columns
\describe{
 \item{Driver Name}
 \item{Permanent Number}
 \item{Nationality}
 \item{DOB}
 \item{Season}
}
}
\source{
\url{http://ergast.com/mrd/methods/drivers/}
}
\usage{
DriverInformation
}
\description{
Driver Information
}
\keyword{datasets}
