% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DriverStandings}
\alias{DriverStandings}
\title{Driver Standings}
\format{
A data.frame with 7 columns
\describe{
 \item{Pos}
 \item{Driver}
 \item{Constructor}
 \item{Points}
 \item{Wins}
 \item{Season}
 \item{Round}
}
}
\source{
\url{http://ergast.com/mrd/methods/standings/}
}
\usage{
DriverStandings
}
\description{
Driver Standings
}
\keyword{datasets}
