% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LapResults}
\alias{LapResults}
\title{Lap time data}
\format{
A data.frame with 6 columns
\describe{
 \item{DriverId}
 \item{Position}
 \item{Time}
 \item{Season}
 \item{Round}
 \item{Lap}
}
}
\source{
\url{http://ergast.com/mrd/methods/laps/}
}
\usage{
LapResults
}
\description{
Lap time data is available from the 1996 season onwards
}
\keyword{datasets}
