% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QualifyingResults}
\alias{QualifyingResults}
\title{List of the qualifying results}
\format{
A data.frame with 9 columns
\describe{
 \item{Pos}
 \item{No}
 \item{Driver}
 \item{Constructor}
 \item{Q1}
 \item{Q2}
 \item{Q3}
 \item{Season}
 \item{Round}
}
}
\source{
\url{https://ergast.com/mrd/methods/qualifying/}
}
\usage{
QualifyingResults
}
\description{
Note: Qualifying results are only fully supported from the 2003 season onwards.
}
\keyword{datasets}
