% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RaceResults}
\alias{RaceResults}
\title{List of the race results}
\format{
A data.frame with 11 columns
\describe{
 \item{Pos}
 \item{No}
 \item{Driver}
 \item{Constructor}
 \item{Laps}
 \item{Grid}
 \item{Time}
 \item{Statis}
 \item{Points}
 \item{Season}
}
}
\source{
\url{http://ergast.com/mrd/methods/results/}
}
\usage{
RaceResults
}
\description{
List of the race results
}
\keyword{datasets}
