% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SeasonList}
\alias{SeasonList}
\title{List of F1 seasons and grands pix}
\format{
A data.frame with 8 columns
\describe{
 \item{Season}
 \item{Round}
 \item{RaceName}
 \item{Date}
 \item{Time}
 \item{Circuit}
 \item{Locality}
 \item{Country}
}
}
\source{
\url{https://ergast.com/mrd/methods/seasons/}
}
\usage{
SeasonList
}
\description{
Data set of F1 seasons and GPs
}
\keyword{datasets}
